Move reticulum tests
