+++ src/test/java/org/qortal/test/network/RNSNetworkTest.java
@@ -1,4 +1,4 @@
-package org.qortal.test.network.reticulum;
+package org.qortal.test.network;
 
 import org.apache.commons.lang3.StringUtils;
 //import org.junit.Before;
